requirements.txt
hydra-core
hydra-colorlog
torch
torchvision
matplotlib